refactor: extract DWC3 gadget configuration into config module
Move DWC3 gadget speed, SUSPHY, and run-stop helpers (plus associated tests) into a dedicated config module, keeping usb.rs focused on early-boot polling and handoff control flow.
